+++ src/components/TransactionsTable/TransactionsTable.tsx
@@ -1,3 +1,5 @@
+import { useMemo } from 'react';
+
 import { MAX_TRANSACTIONS_PAGE_SIZE } from 'appConstants';
 import {
   Pager,
@@ -28,15 +30,26 @@ export const TransactionsTable = ({
       Live Transactions <MethodList />
     </h5>
   ),
-  showDirectionCol = false,
-  showLockedAccounts = false,
-  dataChanged = false,
-  isScResultsTable = false,
+  showDirectionCol,
+  showLockedAccounts,
+  dataChanged,
+  isScResultsTable,
+  hasPauseButton,
+  hasTxPreviewBtn = true,
   isDataReady,
-  inactiveFilters = [TransactionFiltersEnum.isRelayed],
-  hasPauseButton
+  inactiveFilters = [TransactionFiltersEnum.isRelayed]
 }: TransactionTableType) => {
-  const colSpan = showDirectionCol ? 8 : 7;
+  const colSpan = useMemo(() => {
+    let colspan = 7;
+    if (showDirectionCol) {
+      colspan++;
+    }
+    if (hasTxPreviewBtn) {
+      colspan++;
+    }
+
+    return colspan;
+  }, [showDirectionCol, hasTxPreviewBtn]);
 
   return (
     <div
@@ -70,6 +83,7 @@ export const TransactionsTable = ({
                 showLockedAccounts={showLockedAccounts}
                 inactiveFilters={inactiveFilters}
                 hasPauseButton={hasPauseButton}
+                hasTxPreviewBtn={hasTxPreviewBtn}
               />
               <tbody>
                 {isDataReady === undefined && (
@@ -99,6 +113,7 @@ export const TransactionsTable = ({
                             token={token}
                             showDirectionCol={showDirectionCol}
                             showLockedAccounts={showLockedAccounts}
+                            hasTxPreviewBtn={hasTxPreviewBtn}
                           />
                         ))}
                       </>