+++ src/components/TransactionsTable/components/Header/Header.tsx
@@ -14,14 +14,15 @@ import {
 } from 'components';
 import { useIsSovereign } from 'hooks';
 import { transactionsSelector } from 'redux/selectors';
-import { pauseRefresh, resumeRefresh } from 'redux/slices/transactions';
+import { pauseTxRefresh, resumeTxRefresh } from 'redux/slices';
 import { TransactionTableType } from 'types';
 
 export const Header = ({
   showDirectionCol = false,
   address,
   inactiveFilters,
-  hasPauseButton
+  hasPauseButton,
+  hasTxPreviewBtn
 }: TransactionTableType) => {
   const isSovereign = useIsSovereign();
   const { isRefreshPaused } = useSelector(transactionsSelector);
@@ -29,6 +30,7 @@ export const Header = ({
   return (
     <thead>
       <tr>
+        {hasTxPreviewBtn && <th scope='col'></th>}
         <th scope='col'>
           Txn Hash <StatusColumnFilters inactiveFilters={inactiveFilters} />
         </th>
@@ -72,8 +74,8 @@ export const Header = ({
           </div>
           {hasPauseButton && (
             <PauseRefreshButton
-              pauseRefresh={pauseRefresh}
-              resumeRefresh={resumeRefresh}
+              pauseRefresh={pauseTxRefresh}
+              resumeRefresh={resumeTxRefresh}
               isRefreshPaused={isRefreshPaused}
             />
           )}