+++ src/redux/selectors/index.ts
@@ -1,11 +1,11 @@
 export * from './account';
 export * from './nodes';
+export * from './transactions';
 export * from './widgets';
 
 export * from './blocks';
 export * from './collection';
 export * from './economics';
-export * from './events';
 export * from './general';
 export * from './interface';
 export * from './markers';
@@ -19,5 +19,3 @@ export * from './stakeExtra';
 export * from './stats';
 export * from './token';
 export * from './tokenExtra';
-export * from './transactions';
-export * from './transactionsInPool';