+++ src/redux/slices/transactions/transactionsInPool.ts
@@ -38,16 +38,19 @@ export const transactionsInPoolSlice = createSlice({
       state.isDataReady = action.payload.isDataReady;
       state.isWebsocket = action.payload.isWebsocket;
     },
-    pauseRefresh: (state: TransactionInPoolSliceType) => {
+    pauseTxPoolRefresh: (state: TransactionInPoolSliceType) => {
       state.isRefreshPaused = true;
     },
-    resumeRefresh: (state: TransactionInPoolSliceType) => {
+    resumeTxPoolRefresh: (state: TransactionInPoolSliceType) => {
       state.isRefreshPaused = false;
     }
   }
 });
 
-export const { setTransactionsInPool, pauseRefresh, resumeRefresh } =
-  transactionsInPoolSlice.actions;
+export const {
+  setTransactionsInPool,
+  pauseTxPoolRefresh,
+  resumeTxPoolRefresh
+} = transactionsInPoolSlice.actions;
 
 export const transactionsInPoolReducer = transactionsInPoolSlice.reducer;